# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ride-Hailing Platform API
  version: 1.0.0
  description: |
    REST API for riders, drivers, operations staff and admins. Design: [HLD §10](architecture.md#10-api-design),
    [LLD §13](low-level-design.md#13-rest-api).

    - **Authentication**: `Authorization: Bearer <access token>` from `/v1/auth/token` (JWT, 15 min). Sign-in and
      webhook endpoints are public; webhooks are signed.
    - **Idempotency**: commands that change a ride, offer, driver status, payment or rating require an
      `Idempotency-Key` (1–255 visible ASCII characters). A retry with the same key and request returns the original
      response with `Idempotent-Replayed: true`; the same key with a different request gets
      `422 IDEMPOTENCY_KEY_REUSED`; a key still executing gets `409 IDEMPOTENCY_KEY_IN_PROGRESS` with `Retry-After`.
      Keys are kept for 24 hours.
    - **Format**: JSON in snake_case. Money is `{"amount_paise": 25900, "currency": "INR"}`. Timestamps are
      RFC 3339 in UTC. Fields without a value are omitted. Clients must ignore unknown fields.
    - **Errors**: RFC 9457 problem details (`application/problem+json`) with a stable `code` and the `request_id`.
      Any operation may also answer `500 INTERNAL_ERROR` or `503 SERVICE_UNAVAILABLE` (with `Retry-After`); retry
      commands with the same `Idempotency-Key` and backoff.
    - **Visibility**: a resource the caller may not see answers `404`, not `403`.
...

...
servers:
...
security:
  - bearer: []
...
paths:
...
  /v1/riders/me/rides:
    get:
      tags: [Riders]
      operationId: listRiderRides
      summary: Ride history
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': {$ref: '#/components/responses/Rides'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /v1/rides:
    post:
      tags: [Rides]
      operationId: bookRide
      summary: Book a ride from a quote
      description: Answers once the ride is stored in `SEARCHING`; dispatch starts at once.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BookRideRequest'}
      responses:
        '201': {$ref: '#/components/responses/RideCreated'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
  /v1/rides/{ride_id}:
    get:
      tags: [Rides]
      operationId: getRide
      summary: A ride, as seen by the caller
      description: The rider's view includes the PIN and the driver and vehicle; the driver's view includes the rider's first name and rating, never the PIN.
      parameters:
        - $ref: '#/components/parameters/RideId'
      responses:
        '200': {$ref: '#/components/responses/Ride'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/rides/{ride_id}/cancel:
    post:
      tags: [Rides]
      operationId: cancelRide
      summary: Cancel as the rider or the assigned driver
      description: |
        The outcome depends on the caller and the state ([ride lifecycle §3](ride-lifecycle.md#3-transitions)). A driver
        cancelling before arrival returns the ride to `SEARCHING` for the rider. A rider cancelling after the free
        window pays the cancellation fee shown in `cancellation.fee`.
      parameters:
        - $ref: '#/components/parameters/RideId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CancelRideRequest'}
      responses:
        '200': {$ref: '#/components/responses/Ride'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/rides/{ride_id}/arrive:
    post:
      tags: [Rides]
      operationId: markArrived
      summary: The driver is at the pickup
      parameters:
        - $ref: '#/components/parameters/RideId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {$ref: '#/components/responses/Ride'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/rides/{ride_id}/start:
    post:
      tags: [Rides]
      operationId: startTrip
      summary: Start the trip with the rider's PIN
      description: Five wrong PINs lock the start (`409 PIN_LOCKED`). Offline apps send `device_time` (V2).
      parameters:
        - $ref: '#/components/parameters/RideId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StartTripRequest'}
      responses:
        '200': {$ref: '#/components/responses/Ride'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /v1/rides/{ride_id}/complete:
    post:
      tags: [Rides]
      operationId: completeTrip
      summary: End the trip at the drop-off
      description: The fare is the quoted fare. Payment happens afterwards and doesn't change the ride.
      parameters:
        - $ref: '#/components/parameters/RideId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CompleteTripRequest'}
      responses:
        '200': {$ref: '#/components/responses/Ride'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/rides/{ride_id}/no-show:
    post:
      tags: [Rides]
      operationId: markNoShow
      summary: End the ride because the rider didn't come
      description: Allowed 5 minutes after arrival; the rider pays the no-show fee.
      parameters:
        - $ref: '#/components/parameters/RideId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {$ref: '#/components/responses/Ride'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /v1/rides/{ride_id}/receipt:
    get:
      tags: [Rides]
      operationId: getReceipt
      summary: Receipt of a completed ride
      description: The rider's alone, since it shows their payments; a ride not completed is `409 RECEIPT_NOT_AVAILABLE`.
      parameters:
        - $ref: '#/components/parameters/RideId'
      responses:
        '200':
          description: Receipt.
          headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Receipt'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/rides/{ride_id}/route:
    get:
      tags: [Rides]
      operationId: getTripRoute
      x-since: V2
      summary: (V2) The trip route
      description: Points in sequence order, thinned to at most 2,000. Reads by operations are audited.
      parameters:
        - $ref: '#/components/parameters/RideId'
      responses:
        '200':
          description: Route.
          headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TripRoute'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}

  # ------------------------------------------------------------------ Drivers and offers
...
  /v1/drivers/me/rides:
    get:
      tags: [Drivers]
      operationId: listDriverRides
      summary: Trip history
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': {$ref: '#/components/responses/Rides'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /v1/ops/rides:
    get:
      tags: [Operations]
      operationId: listRidesForOps
      summary: Rides by status and city
      parameters:
        - {name: status, in: query, schema: {type: array, items: {$ref: '#/components/schemas/RideStatus'}}, style: form, explode: false}
        - {name: city_id, in: query, schema: {$ref: '#/components/schemas/CityId'}}
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': {$ref: '#/components/responses/Rides'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /v1/ops/rides/{ride_id}/cancel:
    post:
      tags: [Operations]
      operationId: cancelRideAsOps
      summary: Cancel a ride as operations (`CANCELLED_BY_SYSTEM`)
      parameters:
        - $ref: '#/components/parameters/RideId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OpsCancelRequest'}
      responses:
        '200': {$ref: '#/components/responses/Ride'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/v1/auth/token` or `/v1/auth/refresh`.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique per logical command; reuse it only to retry the same request.
      schema: {type: string, minLength: 1, maxLength: 255, pattern: '^[\x21-\x7E]+$'}
    RideId:
      name: ride_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
...
    Cursor:
      name: cursor
      in: query
      description: Opaque; from `next_cursor`.
      schema: {type: string, maxLength: 200}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  headers:
    RequestId:
      required: true
      description: Correlation ID of this request.
      schema: {type: string}
    IdempotentReplayed:
      description: '`true` when this is the stored response of an earlier request with the same key.'
      schema: {type: string, enum: ['true', 'false']}
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: {type: integer, minimum: 1}
  responses:
...
    Ride:
      description: The ride, as seen by the caller.
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Ride'}
    RideCreated:
      description: The ride, in `SEARCHING`.
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Ride'}
    Rides:
      description: Rides, newest first.
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/RideList'}
...
    BadRequest:
      description: '`VALIDATION_FAILED` (see `errors`), `MALFORMED_REQUEST` or `IDEMPOTENCY_KEY_REQUIRED`.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: '`UNAUTHENTICATED`, `CODE_INVALID`, `REFRESH_TOKEN_INVALID` or `WEBHOOK_SIGNATURE_INVALID`.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Forbidden:
      description: '`FORBIDDEN` (wrong role) or `ACCOUNT_DISABLED`.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: '`NOT_FOUND`: unknown, or not visible to the caller.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: |
        The state doesn't allow this now; re-read the resource. Codes: `INVALID_TRANSITION`, `RIDE_REASSIGNED`,
        `OFFER_NO_LONGER_AVAILABLE`, `QUOTE_EXPIRED`, `QUOTE_ALREADY_USED`, `ACTIVE_RIDE_EXISTS`, `DUES_OUTSTANDING`,
        `PIN_LOCKED`, `NO_SHOW_TOO_EARLY`, `DRIVER_NOT_ELIGIBLE`, `DRIVER_HAS_ACTIVE_RIDE`, `PLACES_LIMIT_REACHED`,
        `NO_DUES`, `CHARGE_NOT_REFUNDABLE`, `RATING_NOT_OPEN`, `ALREADY_RATED`, `RECEIPT_NOT_AVAILABLE`,
        `VERSION_CONFLICT` (with `current_version`), `ALREADY_EXISTS`, `FLAG_ALREADY_RESOLVED`, or
        `IDEMPOTENCY_KEY_IN_PROGRESS` (retry with the same key after `Retry-After`).
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unprocessable:
      description: |
        Don't retry unchanged. Codes: `IDEMPOTENCY_KEY_REUSED`, `OUTSIDE_SERVICE_AREA`, `CATEGORY_NOT_AVAILABLE`,
        `ROUTE_NOT_FOUND`, `PAYMENT_METHOD_INVALID`, `WRONG_PIN` (with `attempts_left`), `REFUND_EXCEEDS_CHARGE`,
        `FEE_EXCEEDS_RULE`, `INVALID_GEOMETRY`, `RULE_EFFECTIVE_IN_PAST`.
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    TooManyRequests:
      description: '`RATE_LIMITED` or `CODE_ATTEMPTS_EXCEEDED`. Not processed, and the `Idempotency-Key` was not consumed.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Uuid: {type: string, format: uuid}
    Timestamp: {type: string, format: date-time}
    CityId: {type: string, pattern: '^[a-z]{3,8}$', examples: [blr]}
    CategoryCode: {type: string, pattern: '^[A-Z][A-Z0-9_]{1,19}$', examples: [MINI]}
...
    Money:
      type: object
      required: [amount_paise, currency]
      properties:
        amount_paise: {type: integer, format: int64, description: 'Paise; negative only in earnings adjustments.'}
        currency: {type: string, pattern: '^[A-Z]{3}$', examples: [INR]}
    GeoPoint:
      type: object
      required: [lat, lon]
      properties:
        lat: {type: number, minimum: -90, maximum: 90}
        lon: {type: number, minimum: -180, maximum: 180}
    RatingSummary:
      type: object
      required: [count]
      properties:
        average: {type: number, minimum: 1, maximum: 5, description: Average of the latest 100 ratings; absent with no ratings.}
        count: {type: integer, minimum: 0, maximum: 100}
    Problem:
      type: object
      required: [title, status, code, request_id]
      properties:
        type: {type: string, default: about:blank}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string, description: Stable machine-readable code (LLD §13.2).}
        request_id: {type: string}
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
        current_status: {type: string}
        current_version: {type: integer}
        attempts_left: {type: integer, minimum: 0}
        available_at: {$ref: '#/components/schemas/Timestamp'}
        dues: {$ref: '#/components/schemas/Money'}
    RideStatus:
      type: string
      enum: [SEARCHING, DRIVER_ASSIGNED, DRIVER_ARRIVED, IN_TRIP, COMPLETED, CANCELLED_BY_RIDER, CANCELLED_BY_DRIVER, CANCELLED_BY_SYSTEM, DRIVER_NOT_FOUND]
...
    ChargeStatus: {type: string, enum: [PENDING, SUCCEEDED, FAILED, UNKNOWN]}
    ChargePurpose: {type: string, enum: [FARE, CANCELLATION_FEE, NO_SHOW_FEE]}
    MethodType: {type: string, enum: [CASH, CARD, UPI]}
...
    ChargeSummary:
      type: object
      required: [id, ride_id, purpose, amount, status, method_type, created_at]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        ride_id: {$ref: '#/components/schemas/Uuid'}
        purpose: {$ref: '#/components/schemas/ChargePurpose'}
        amount: {$ref: '#/components/schemas/Money'}
        status: {$ref: '#/components/schemas/ChargeStatus'}
        method_type: {$ref: '#/components/schemas/MethodType'}
        failure_code: {type: string}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
...
    FareBreakdown:
      type: object
      required: [base, distance, time, surge, minimum_topup, booking_fee, tax, rounding, total]
      properties:
        base: {$ref: '#/components/schemas/Money'}
        distance: {$ref: '#/components/schemas/Money'}
        time: {$ref: '#/components/schemas/Money'}
        surge: {$ref: '#/components/schemas/Money'}
        minimum_topup: {$ref: '#/components/schemas/Money'}
        booking_fee: {$ref: '#/components/schemas/Money'}
        tax: {$ref: '#/components/schemas/Money'}
        rounding: {$ref: '#/components/schemas/Money'}
        total: {$ref: '#/components/schemas/Money'}
...
    BookRideRequest:
      type: object
      required: [quote_id]
      properties:
        quote_id: {$ref: '#/components/schemas/Uuid'}
        payment_method_id: {$ref: '#/components/schemas/Uuid', description: The default method when absent.}
    PersonSummary:
      type: object
      required: [id, first_name]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        first_name: {type: string}
        rating: {$ref: '#/components/schemas/RatingSummary'}
    VehicleSummary:
      type: object
      required: [id, category, make, model, colour, plate]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        category: {$ref: '#/components/schemas/CategoryCode'}
        make: {type: string}
        model: {type: string}
        colour: {type: string}
        plate: {type: string}
    Cancellation:
      type: object
      required: [cancelled_by]
      properties:
        cancelled_by: {type: string, enum: [RIDER, DRIVER, SYSTEM]}
        reason: {type: string}
        fee:
          type: object
          required: [purpose, amount]
          properties:
            purpose: {type: string, enum: [CANCELLATION_FEE, NO_SHOW_FEE]}
            amount: {$ref: '#/components/schemas/Money'}
    Ride:
      type: object
      required: [id, status, version, city_id, category, pickup, dropoff, fare, payment_method, requested_at]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        status: {$ref: '#/components/schemas/RideStatus'}
        version: {type: integer, minimum: 0}
        city_id: {$ref: '#/components/schemas/CityId'}
        category: {$ref: '#/components/schemas/CategoryCode'}
        pickup: {$ref: '#/components/schemas/GeoPoint'}
        dropoff: {$ref: '#/components/schemas/GeoPoint'}
        fare: {$ref: '#/components/schemas/Money'}
        payment_method:
          type: object
          required: [id, type]
          properties:
            id: {$ref: '#/components/schemas/Uuid'}
            type: {$ref: '#/components/schemas/MethodType'}
        pin: {type: string, pattern: '^[0-9]{4}$', description: Rider's view only, from assignment until the trip starts.}
        promised_pickup_eta_s: {type: integer, minimum: 0}
        driver: {$ref: '#/components/schemas/PersonSummary'}
        vehicle: {$ref: '#/components/schemas/VehicleSummary'}
        rider: {$ref: '#/components/schemas/PersonSummary'}
        cancellation: {$ref: '#/components/schemas/Cancellation'}
        earnings: {$ref: '#/components/schemas/TripEarnings'}
        requested_at: {$ref: '#/components/schemas/Timestamp'}
        assigned_at: {$ref: '#/components/schemas/Timestamp'}
        arrived_at: {$ref: '#/components/schemas/Timestamp'}
        started_at: {$ref: '#/components/schemas/Timestamp'}
        completed_at: {$ref: '#/components/schemas/Timestamp'}
        ended_at: {$ref: '#/components/schemas/Timestamp'}
    RideList:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Ride'}}
        next_cursor: {type: string}
    TripEarnings:
      type: object
      description: The driver's earnings from a completed ride, in the driver's and operations' views only (FR-D4).
      required: [gross, commission, net, cash_collected]
      properties:
        gross: {$ref: '#/components/schemas/Money'}
        commission: {$ref: '#/components/schemas/Money'}
        net: {$ref: '#/components/schemas/Money'}
        cash_collected: {$ref: '#/components/schemas/Money'}
    CancelRideRequest:
      type: object
      properties:
        reason: {type: string, maxLength: 200}
    StartTripRequest:
      type: object
      required: [pin]
      properties:
        pin: {type: string, pattern: '^[0-9]{4}$'}
        device_time: {$ref: '#/components/schemas/Timestamp'}
    CompleteTripRequest:
      type: object
      properties:
        device_time: {$ref: '#/components/schemas/Timestamp'}
...
    Receipt:
      type: object
      required: [ride_id, city_id, category, pickup, dropoff, distance_m, duration_s, fare, payment, completed_at]
      properties:
        ride_id: {$ref: '#/components/schemas/Uuid'}
        city_id: {$ref: '#/components/schemas/CityId'}
        category: {$ref: '#/components/schemas/CategoryCode'}
        pickup: {$ref: '#/components/schemas/GeoPoint'}
        dropoff: {$ref: '#/components/schemas/GeoPoint'}
        distance_m: {type: integer, minimum: 0, description: Quoted route distance.}
        duration_s: {type: integer, minimum: 0}
        fare: {$ref: '#/components/schemas/FareBreakdown'}
        payment:
          type: object
          required: [method_type, charges]
          properties:
            method_type: {$ref: '#/components/schemas/MethodType'}
            charges: {type: array, items: {$ref: '#/components/schemas/ChargeSummary'}}
        driver: {$ref: '#/components/schemas/PersonSummary'}
        vehicle: {$ref: '#/components/schemas/VehicleSummary'}
        started_at: {$ref: '#/components/schemas/Timestamp'}
        completed_at: {$ref: '#/components/schemas/Timestamp'}
    TripRoute:
      type: object
      required: [ride_id, points, thinned]
      properties:
        ride_id: {$ref: '#/components/schemas/Uuid'}
        thinned: {type: boolean}
        points:
          type: array
          maxItems: 2000
          items:
            type: object
            required: [seq, lat, lon, received_at]
            properties:
              seq: {type: integer, format: int64}
              lat: {type: number}
              lon: {type: number}
              received_at: {$ref: '#/components/schemas/Timestamp'}

    # ---------------------------------------------------------------- drivers and offers
...
    OpsCancelRequest:
      type: object
      required: [reason]
      properties:
        reason: {type: string, minLength: 1, maxLength: 500}
        fee:
          type: object
          required: [purpose, amount_paise]
          properties:
            purpose: {type: string, enum: [CANCELLATION_FEE, NO_SHOW_FEE]}
            amount_paise: {type: integer, format: int64, minimum: 1, description: At most the fee rule's amount.}
...